reorder homepage sections
