+++ src/containers/Homepage/HomeHero.module.scss
@@ -48,7 +48,6 @@
   }
 }
 .articles_container {
-  margin-top: 104px;
 }
 .articles {
   border-top: 1px solid var(--color-gray-vlight);
