+++ src/containers/Homepage/Homepage.tsx
@@ -9,6 +9,7 @@ import { NFTArticle } from "../../types/entities/Article"
 import { LiveMintingEvent } from "../../types/entities/LiveMinting"
 import { HomeEvents } from "./HomeEvents"
 import { HomeIncoming } from "./HomeIncoming"
+import { HomeDocs } from "./HomeDocs"
 
 interface HomepageProps {
   generativeTokens: GenerativeToken[]
@@ -27,15 +28,13 @@ const _Homepage = ({
 }: HomepageProps) => {
   return (
     <div className={style.container}>
-      <HomeHero
-        articles={articles}
-        randomGenerativeToken={randomGenerativeToken}
-      />
+      <HomeHero randomGenerativeToken={randomGenerativeToken} />
+      <HomeMarketplace />
       {events.length > 0 && <HomeEvents events={events} />}
+      <HomeDocs articles={articles} />
       {incomingTokens.length > 0 && (
         <HomeIncoming generativeTokens={incomingTokens} />
       )}
-      <HomeMarketplace />
       <HomeExplore generativeTokens={generativeTokens} />
       <HomeGetStarted />
     </div>
